Add flag to clojure cli to fully support XDG
The flag could be set at installation time the way the changes in the
`linux-install.sh` script suggest.

Set or not, if the user defines the environment variables
`XDG_CONFIG_HOME` or `XDG_CACHE_HOME`, those values would be used,
exactly as the current version of the Clojure CLI does.

If the clojure cli is installed to "fully support XDG", the default
base XDG directories would be used unless the mentioned environment
variables are set or if the environment variables CLJ_CONFIG and
CLJ_CACHE are defined, which still have the top priority (not changing
any original behaviour).